+++ scripts/lib/utilities/workflowUtils.js
@@ -282,6 +282,15 @@ async function specialItemUse(item, targets, sourceFeature, {activity} = {}) {
     }
     await genericUtils.remove(effect);
 }
+async function updateTargets(workflow, targets, {userId} = {}) {
+    workflow.targets = new Set(targets);
+    await genericUtils.updateTargets(targets, userId);
+}
+async function removeTargets(workflow, tokens, {userId} = {}) {
+    let targets = Array.from(workflow.targets).filter(token => !tokens.includes(token));
+    workflow.targets = new Set(targets);
+    await genericUtils.updateTargets(targets, userId);
+}
 export let workflowUtils = {
     bonusDamage,
     bonusAttack,
@@ -302,5 +311,7 @@ export let workflowUtils = {
     modifyDamageAppliedFlat,
     getCastLevel,
     syntheticActivityDataRoll,
-    specialItemUse
+    specialItemUse,
+    updateTargets,
+    removeTargets
 };