+++ scripts/macros/2024/classFeatures/rogue/arcaneTrickster/versatileTrickster.js
@@ -1,0 +1,5 @@
+export let versatileTrickster = {
+    name: 'Versatile Trickster',
+    version: '1.3.53',
+    rules: 'modern'
+};