+++ scripts/macros/2024/spells/massCureWounds.js
@@ -1,10 +1,10 @@
-import {dialogUtils, genericUtils} from '../../../utils.js';
+import {dialogUtils, genericUtils, workflowUtils} from '../../../utils.js';
 async function early({workflow}) {
     if (!workflow.targets.size) return;
     let sourceDisposition = workflow.token.document.disposition;
     let targetTokens = Array.from(workflow.targets.filter(i => i.document.disposition === sourceDisposition));
     if (!targetTokens.length || targetTokens.length <= 6) {
-        genericUtils.updateTargets(targetTokens);
+        await workflowUtils.updateTargets(workflow, targetTokens);
         return;
     }
     let [selection] = await dialogUtils.selectTargetDialog(workflow.item.name, 'CHRISPREMADES.Macros.MassCureWounds.Select', targetTokens, {type: 'multiple', maxAmount: 6});
@@ -14,7 +14,7 @@ async function early({workflow}) {
     } else {
         newTargets = selection ?? [];
     }
-    genericUtils.updateTargets(newTargets);
+    await workflowUtils.updateTargets(workflow, newTargets);
 }
 export let massCureWounds = {
     name: 'Mass Cure Wounds',